Create Rust library for painting Bytecode
A new Rust library named 'bytekot-painter' has been created that renders Bytecode using the ANTLR lexer 'JBytecodeLexer'. The library, using Skia for rendering, can perform lexing and token categorization which are then styled and laid out using Skia text layout. The output is then rasterized and saved as a PNG image. The library can be consumed as a static library.