Update library API and docker version tag in readme file
This commit updates the API in the readme file, changing the `paint` function to return a structure of data instead of a path and introduces the `free_image_data` function. It also updates the docker image version tag and changes the generated library type from static to dynamic for both Windows and Linux versions.