Update project documentation and Docker image version
This commit does several modifications in the README files, both in English and Russian versions. It changes the description of how the 'paint' method functions, removing the part about returning a C-style string. It also updates the Docker image version being used from v1.0.1 to v1.1.3. Steps on building the library are updated to reflect the shift from a static library to a dynamic one. Finally, a new API section is added to explain how the 'paint' and 'free_image_data' functions work.